LPS-57153 move
