+++ portal-impl/src/com/liferay/portal/service/impl/ResourceLocalServiceImpl.java
@@ -33,10 +33,10 @@ import com.liferay.portal.model.impl.ResourceImpl;
 import com.liferay.portal.security.permission.PermissionCacheUtil;
 import com.liferay.portal.security.permission.PermissionThreadLocal;
 import com.liferay.portal.security.permission.ResourceActionsUtil;
-import com.liferay.portal.service.ModelPermissions;
-import com.liferay.portal.service.ModelPermissionsFactory;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.base.ResourceLocalServiceBaseImpl;
+import com.liferay.portal.service.permission.ModelPermissions;
+import com.liferay.portal.service.permission.ModelPermissionsFactory;
 import com.liferay.portal.util.ResourcePermissionsThreadLocal;
 
 import java.util.Arrays;
